+++ testData/org/elixir_lang/heex/reference/heex_component/use_list_of_quotes/component_library.ex
@@ -1,0 +1,16 @@
+defmodule MyComponentLibrary do
+  defmacro __using__(_opts) do
+    [
+      quote do
+        import MyComponentLibrary.Helpers
+      end,
+      quote do
+        import MyComponentLibrary
+      end
+    ]
+  end
+
+  def link(assigns) do
+    ~H""
+  end
+end