+++ testData/org/elixir_lang/heex/reference/heex_component/use_match_last/component_library.ex
@@ -1,0 +1,12 @@
+defmodule MyComponentLibrary do
+  defmacro __using__(_opts) do
+    imports =
+      quote do
+        import MyComponentLibrary
+      end
+  end
+
+  def link(assigns) do
+    ~H""
+  end
+end